Setting up the DATA base population from data
translated the fast.py into JS and expanded it.

Added npm packages:
-unzipper
-pandas
-jsonfile

When a file is uplaoded the zips are unloaded, the data is stripped,
added to a JSON then the JSON is read and quried into the database

Curently creating the quries to upload the data.

Also created markdown for personal notes.